+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - DMA flowmeters & connecs endpoints
 - Tests
 
+### Changed
+
+- DMA endpoints prefix
+
 ## [0.6.0] - 2026-01-29
 
 ### Added