+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2026-02-10
+
 ### Added
 
 - DMA connecs endpoints
@@ -148,7 +150,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic test with pytest.
 - Basic CI workflow.
 
-[unreleased]: https://github.com/giswater/giswater-api/compare/v0.6.0...main
+[unreleased]: https://github.com/giswater/giswater-api/compare/v0.7.0...main
+[0.7.0]: https://github.com/giswater/giswater-api/compare/v0.6.0...0.7.0
 [0.6.0]: https://github.com/giswater/giswater-api/compare/v0.5.0...0.6.0
 [0.5.0]: https://github.com/giswater/giswater-api/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/giswater/giswater-api/compare/v0.3.4...v0.4.0